+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Only deploy documentation on keithgroup repo.
+- Correct dataSet Rset_info documentation.
+
 ## [0.0.1] - 2022-02-24
 
 - Initial release!